requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0

# Google Cloud / Gemini
google-genai==1.0.0
google-cloud-aiplatform>=1.38.0

# Datadog
ddtrace>=2.4.0
datadog-api-client>=2.20.0
datadog>=0.49.0

# Utilities
tenacity>=8.2.3
rich>=13.7.0
numpy>=1.26.0
